Added Login modal
For some reason the Javascript isn't working - regi.js located in the
javascript files.

Also the second modal for the contact form has also gone retarded
because of the css settings for the login modal as their both sharing
the same classes.

Login form needs to be complete with a register/sign up option on the
login form preferably.

some bugs with button selection also needs to be fixed.
(some buttons will have selection lines around them after they've been
selected.

the project submission form needs to be completed, I've found a really
good template here (obviously the background will be changed also the
form will be wider so all options fit on one screen)
http://bootsnipp.com/snippets/featured/bootstrap-form-theme

asp.net OR php needs to be used to return the information given within
the forms either by email or to a database such as SQL which we can rent
for about £3 a month i guess? need to look into it more.

With the login and members area and whatnot, i think we should postpone
it for now but start working on the basics because we need a database
connection etc.

These tasts are for the both of us btw :)